# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My-brand API",
    "description": "API",
    "version": "1.0.0"
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "name": "DEV"
    },
    {
      "url": "https://mynewbrandapi.herokuapp.com",
      "name": "PROD"
    }
  ],
  "paths": {
...
    "/api/v1/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get one article",
        "description": "Get one article",
        "operationId": "getoneArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": ["Articles"],
        "summary": "Update article",
        "description": "Update article",
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true
          },
          {
            "name": "title",
            "in": "formData",
            "description": "title",
            "required": true
          },
          {
            "name": "content",
            "in": "formData",
            "description": "content",
            "required": true
          },
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "description": "image",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["Articles"],
        "summary": "Delete one article",
        "description": "Delete one article",
        "operationId": "deleteoneArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
         "security": [
        {
          "Bearer": []
        }
      ]
      }
     
    },
...
  },
...
}